Python blocks can now save to a file

+++ test/testFlowDocument.cpp
@@ -3,7 +3,7 @@
 #include <filesystem>
 #include <Flow/FlowDocument.h>
 
-TEST_CASE("FlowDocument no name errors.") {
+TEST_CASE("FlowDocument file path errors.") {
 	Flow::FlowDocument doc;
 	SUBCASE("Throws when trying to open a document with no name.") {
 		CHECK_THROWS(doc.open());
@@ -13,17 +13,16 @@ TEST_CASE("FlowDocument no name errors.") {
 	}
 }
 
-TEST_CASE("Default generated FlowDocument can actually produce an output file.") {
+TEST_CASE("Default generated FlowDocument can actually produce a readable output file.") {
 	auto path = std::filesystem::temp_directory_path() / "TestFlowDocument.efd";
 	// If the file already exist, this test would be pretty meaningless
-	if (std::filesystem::exists(path)) {
-		std::filesystem::remove(path);
-	}
+	if (std::filesystem::exists(path)) {std::filesystem::remove(path);}
+	CHECK(!std::filesystem::exists(path));
+
 	Flow::FlowDocument testDoc(path);
 	CHECK(testDoc.save());
-	bool fileExist = std::filesystem::exists(path);
-	CHECK(fileExist);
-	if (fileExist) {
-		std::filesystem::remove(path); //Clean up
-	}
+	testDoc.close();
+	CHECK(std::filesystem::exists(path));
+	CHECK(testDoc.open());
+	testDoc.close();
 }
